+++ packages/ui/src/components/organisms/Form/DynamicForm/fields/EntityFieldGroup/types/index.ts
@@ -1,5 +1,6 @@
 export interface IEntity {
   ballerineEntityId?: string;
+  variant: 'director' | 'ubo';
   __id?: string;
   __isGeneratedAutomatically?: boolean;
 }